+++ src/routes/graphql/queries/member-type.ts
@@ -1,8 +1,7 @@
-import { GraphQLList } from 'graphql/index.js';
-import { GraphQLFieldConfig } from 'graphql/type/definition.js';
+import { GraphQLFieldConfig, GraphQLList } from 'graphql';
 import { GraphQLContext } from '../types/main.js';
-import { getAllMemberTypes } from '../resolvers/member-type.js';
 import { MemberTypeType } from '../types/member-type.js';
+import { getAllMemberTypes } from '../resolvers/member-type.js';
 
 const MemberTypesQuery: GraphQLFieldConfig<undefined, GraphQLContext> = {
   description: 'Get all member types',
