+++ src/routes/graphql/queries/post.ts
@@ -1,5 +1,4 @@
-import { GraphQLList } from 'graphql/index.js';
-import { GraphQLFieldConfig } from 'graphql/type/definition.js';
+import { GraphQLFieldConfig, GraphQLList } from 'graphql';
 import { GraphQLContext } from '../types/main.js';
 import { PostType } from '../types/post.js';
 import { getAllPosts } from '../resolvers/post.js';
