+++ src/routes/graphql/queries/profile.ts
@@ -1,5 +1,4 @@
-import { GraphQLList } from 'graphql/index.js';
-import { GraphQLFieldConfig } from 'graphql/type/definition.js';
+import { GraphQLFieldConfig, GraphQLList } from 'graphql';
 import { GraphQLContext } from '../types/main.js';
 import { ProfileType } from '../types/profile.js';
 import { getAllProfiles } from '../resolvers/profile.js';
