+++ src/routes/graphql/queries/user.ts
@@ -1,8 +1,7 @@
-import { GraphQLList } from 'graphql/index.js';
+import { GraphQLFieldConfig, GraphQLList } from 'graphql';
+import { GraphQLContext } from '../types/main.js';
 import { UserType } from '../types/user.js';
 import { getAllUsers } from '../resolvers/user.js';
-import { GraphQLFieldConfig } from 'graphql/type/definition.js';
-import { GraphQLContext } from '../types/main.js';
 
 const UsersQuery: GraphQLFieldConfig<undefined, GraphQLContext> = {
   description: 'Get all users',
