+++ src/routes/graphql/resolvers/profile.ts
@@ -1,5 +1,5 @@
+import { GraphQLFieldResolver } from 'graphql';
 import { GraphQLContext } from '../types/main.js';
-import { GraphQLFieldResolver } from 'graphql/type/index.js';
 
 export const getAllProfiles: GraphQLFieldResolver<undefined, GraphQLContext, object> = (
   _,
