+++ src/routes/graphql/types/user.ts
@@ -1,6 +1,6 @@
 import { GraphQLFloat, GraphQLNonNull, GraphQLObjectType, GraphQLString } from 'graphql';
-import { UUIDType } from './uuid.js';
 import { GraphQLContext } from './main.js';
+import { UUIDType } from './uuid.js';
 
 export interface IUser {
   id: string;
